feat: remove babylon preview and make unity default
